Refactoring test hooks
